update superMemo code, mainly remove unecessary files, update file name, and add type (ts)
